fix: resolve .js imports
