+++ .env.example
@@ -1,2 +1,3 @@
 MONGODB_URI=mongodb://localhost:27017/your-database-name
 PORT=3000
+DEPLOY_URL=http://localhost:3000
