+++ src/shared/infrastructure/http/server/middlewares/errors/ServerError/ServerError.ts
